api/pyproject.toml
[tool.poetry]
name = "demo-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.103.2"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
sqlalchemy = "^2.0.21"
aiomysql = "^0.2.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
mysql-connector-python = "^8.2.0"
mangum = "^0.17.0"
sqlalchemy-utils = "^0.41.1"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.21.1"
aiosqlite = "^0.19.0"
httpx = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 119

[tool.black]
line-length = 119
